LFLT-480 Update dependencies of Leaflet stack components  - Changed CircleCI build image version  - Changed CircleCI Docker build environment version  - Updated Leaflet Stack Base BOM version to 2.0-r2208.1  - Fixed an H2 incompatibility issue  - Bumped version to 1.6.0-dev

+++ core/src/main/resources/failover_init.sql
@@ -1,13 +1,13 @@
 create table cbfs_status_tracking (
-  id int(11) auto_increment not null primary key,
+  id bigint auto_increment not null primary key,
   created_at datetime not null,
   status varchar(20) not null,
   parameter varchar(20) default null
 );
 
 create table cbfs_entry_page (
-  page_number int(11) not null,
-  category_id int(11) default null,
+  page_number bigint not null,
+  category_id bigint default null,
   link_list text,
   content text
 );
@@ -23,6 +23,6 @@ create table cbfs_document (
 );
 
 create table cbfs_category (
-  id int(11) not null primary key,
+  id bigint not null primary key,
   title varchar(255)
 )
